♻️ refactor(rollup): move shared test fixtures out of the test file
